Lots of work on both the algorithm and the code

+++ src/Item.java
@@ -1,9 +1,9 @@
 public class Item {
 	String itemName;
 	int weight;
-	double value;
+	int value;
 
-	public Item(String itemName, int weight, double value) {
+	public Item(String itemName, int weight, int value) {
 		this.itemName = itemName;
 		this.weight = weight;
 		this.value = value;
@@ -25,18 +25,18 @@ public class Item {
 		this.weight = weight;
 	}
 
-	public double getValue() {
+	public int getValue() {
 		return value;
 	}
 
-	public void setValue(double value) {
+	public void setValue(int value) {
 		this.value = value;
 	}
 
 	@Override
 	public String toString() {
 		String output = "";
-		output = output + String.format("%s %d %.0f", itemName, weight, value);
+		output = output + String.format("%s %d %d", itemName, weight, value);
 		return output;
 	}
 }
